feat: Add classroom assignment components and functionality
- Implemented SelectedAulaInfo component to display selected classroom information.
- Created TicMaterialSection component for managing TIC materials in classroom assignments.
- Developed CrearRecursos and EditarRecursos components for creating and editing TIC resources.
- Added RecursosCard and RecursosHeader components for displaying and managing resources.
- Introduced RecursosList component to filter and list resources.
- Implemented form handling logic in useClassroomFormHandlers for classroom assignment.
- Defined validation schema for classroom assignments in schemasAulas.
