+++ src/app/App.jsx
@@ -1,4 +1,4 @@
-import AppRoutes from "@routes/AppRoutes";
+import AppRoutes from "@/app/(routes)/AppRoutes";
 
 function App() {
   return (
