+++ src/app/components/Recursos/RecursosCard.jsx
